+++ UI/InterruptFrame.lua
@@ -148,8 +148,7 @@ function MP.InterruptFrame:UpdateVisibility()
         return
     end
 
-    local inInstance, instanceType = IsInInstance()
-    local shouldShow = MP:IsInMythicPlus() or instanceType == "party" or instanceType == "raid"
+    local shouldShow = MP:ShouldShowHUD()
 
     if shouldShow then
         self.frame:Show()